requirements.txt
jinja2
django-jinja

pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-tldr==0.2.4
pytest-xdist==2.5.0
